fix: remove parameters from new issue page and move form logic from app.jsx to new.jsx
